Move aggregator to signature-aggregator directory
