+++ signature-aggregator/aggregator/aggregator.go
